refactor: extract core/index.tsx into 7 internal plugins
Split the monolithic core/index.tsx (~5000 lines) into a plugin
architecture with 7 focused plugins communicating via SharedPluginApi
and interceptor chain:

- keyboard-plugin: hold activation, keydown/keyup handling, Escape
- pointer-plugin: mouse/touch tracking, element detection, drag, copy
- navigation-plugin: arrow key navigation, action cycling
- toolbar-plugin: toolbar state, activation/deactivation, freeze effects
- menus-plugin: context menu, comments dropdown
- prompt-plugin: prompt mode, agent integration
- copy-pipeline: copy execution, clipboard, feedback labels

Also fixes:
- Set wasActivatedByToggle during keyboard hold activation so toggle-mode
  semantics (Escape to dismiss, modifier release ignored) work correctly
- Add missing pointermove/pointerdown/pointerup event dispatchers to the
  interceptor chain (selection box rendering depended on these)
- Fix flaky freeze-updates test by replacing fixed timeout with
  waitForFunction for element count assertion
